%% -*- Coding: utf-8 -*-
\encoding{utf-8}
\name{pal_CUFF}
\alias{pal_CUFF}
\title{Palette de couleurs}
\description{
  A 10 color palette.
}
\usage{
pal_CUFF(n = 10, pal = "CUFF")
}
\arguments{
  \item{n}{Integer indicating the number of color desired (1-10)}
  \item{pal}{The only value possible for now is CUFF}
}
\value{
  Returns a vector of color.  
}

\author{Charles-\enc{Édouard}{Edouard} \enc{Giguère}{Giguere} }
\examples{
pal_CUFF(3)
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ palette }
